// src/interfaces/Destino.ts
export enum Destino {
  MEXICO = 0,
  NORTEAMERICA_1 = 1,
  NORTEAMERICA_2 = 2,
  AMERICA_CENTRAL_Y_EL_CARIBE = 3,
  SUDAMERICA_NORTE = 4,
  SUDAMERICA_SUR = 5,
  NORESTE_DE_ASIA = 6,
  SUDESTE_DE_ASIA = 7,
  EUROPA = 8,
  NORTE_DE_AFRICA = 9,
  MEDIO_ORIENTE = 10,
  AFRICA = 11,
  HAWAI = 12,
  SUDOESTE_DE_ASIA = 13,
  AUSTRALIA_NUEVA_ZELANDA_Y_PACIFICO_SUR = 14
}